# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  realworld-endpoints_internal_dto.CommentDTO:
    properties:
      author:
        $ref: '#/definitions/realworld-endpoints_internal_dto.ProfileDTO'
      body:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  realworld-endpoints_internal_dto.CommentResponse:
    properties:
      comment:
        $ref: '#/definitions/realworld-endpoints_internal_dto.CommentDTO'
    type: object
  realworld-endpoints_internal_dto.CommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/realworld-endpoints_internal_dto.CommentDTO'
        type: array
    type: object
  realworld-endpoints_internal_dto.CreateCommentInput:
    properties:
      body:
        type: string
    type: object
  realworld-endpoints_internal_dto.CreateCommentRequest:
    properties:
      comment:
        $ref: '#/definitions/realworld-endpoints_internal_dto.CreateCommentInput'
    type: object
  realworld-endpoints_internal_dto.ProfileDTO:
    properties:
      bio:
        type: string
      following:
        type: boolean
      image:
        type: string
      username:
        type: string
    type: object
  realworld-endpoints_internal_dto.ProfileResponse:
    properties:
      profile:
        $ref: '#/definitions/realworld-endpoints_internal_dto.ProfileDTO'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API implementing the RealWorld spec built with Go, Echo Framework,
    GORM, and PostgreSQL.
  title: RealWorld Conduit API
  version: "1.0"
paths:
...
  /articles/{slug}/comments:
    get:
      description: Get all comments for an article identified by slug
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.CommentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get comments for an article
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a new comment on an article
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Comment payload
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/realworld-endpoints_internal_dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add comment to an article
      tags:
      - Comments
  /articles/{slug}/comments/{id}:
    delete:
      description: Delete a comment (only the author of the comment can delete it)
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
...
  /profiles/{username}:
    get:
      description: Get a user's public profile by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get user profile
      tags:
      - Profiles
  /profiles/{username}/follow:
    delete:
      description: Unfollow a user by username
      parameters:
      - description: Username to unfollow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Unfollow a user
      tags:
      - Profiles
    post:
      description: Follow a user by username
      parameters:
      - description: Username to follow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Follow a user
      tags:
      - Profiles
...
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT Token. Format: "Token {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"